Plugins: categorize linux.netfilter as a malware plugin
